+++ src/sys/socket/sockopt.rs
@@ -148,6 +148,10 @@ macro_rules! sockopt_impl {
         sockopt_impl!(SetOnly, $name, $level, $flag, usize, SetUsize);
     };
 
+    (SetOnly, $name:ident, $level:path, $flag:path, Vec<u8>) => {
+        sockopt_impl!(SetOnly, $name, $level, $flag, Vec<u8>, SetBytes);
+    };
+
     (Both, $name:ident, $level:path, $flag:path, bool) => {
         sockopt_impl!(Both, $name, $level, $flag, bool, GetBool, SetBool);
     };
@@ -219,6 +223,7 @@ cfg_if! {
     if #[cfg(any(target_os = "android", target_os = "linux"))] {
         sockopt_impl!(SetOnly, Ipv6AddMembership, libc::IPPROTO_IPV6, libc::IPV6_ADD_MEMBERSHIP, super::Ipv6MembershipRequest);
         sockopt_impl!(SetOnly, Ipv6DropMembership, libc::IPPROTO_IPV6, libc::IPV6_DROP_MEMBERSHIP, super::Ipv6MembershipRequest);
+        sockopt_impl!(SetOnly, AlgSetKey, libc::SOL_ALG, libc::ALG_SET_KEY, Vec<u8>);
     } else if #[cfg(any(target_os = "dragonfly",
                         target_os = "freebsd",
                         target_os = "ios",
@@ -367,6 +372,24 @@ unsafe impl<T> Get<T> for GetStruct<T> {
     }
 }
 
+/// Setter for bytes.
+struct SetBytes<'a> {
+    ptr: &'a Vec<u8>,
+}
+
+unsafe impl<'a> Set<'a, Vec<u8>> for SetBytes<'a> {
+    fn new(ptr: &'a Vec<u8>) -> SetBytes<'a> {
+        SetBytes { ptr }
+    }
+
+    fn ffi_ptr(&self) -> *const c_void {
+        self.ptr.as_slice() as *const _ as *const c_void
+    }
+
+    fn ffi_len(&self) -> socklen_t {
+        self.ptr.len() as socklen_t
+    }
+}
 /// Setter for an arbitrary `struct`.
 struct SetStruct<'a, T: 'static> {
     ptr: &'a T,